@@ -61,8 +61,9 @@
     "json-loader": "^0.5.4",
     "postcss": "^5.2.11",
     "postcss-loader": "^1.2.2",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
     "standard": "^8.6.0",
     "standard-react": "^4.0.0",
     "style-loader": "^0.13.1",
