@@ -1,12 +1,28 @@
-import React, {PropTypes} from 'react'
+import React from 'react'
 import ReactDOMServer from 'react-dom/server'
 import TestshotWindow from './components/TestshotWindow'
 import ScenarioView from './components/ScenarioView'
+
+let PropTypes
+try {
+  PropTypes = require('prop-types')
+} catch (e) {
+  // Ignore optional peer dependency
+}
 
 const names = []
 const data = []
 var currentContext = null
 
+/**
+ * Create a scenario.
+ * @param {String} name - name of the scenario
+ * @param {React.Component} type - component to create a scenario from
+ * @param {Object} [options]
+ * @param {Boolean} [options.css] - enable CSS diff
+ * @param {Boolean} [options.screenshot] - enable CSS and screenshot diff.
+ *   When true, ignore the value of `options.css`
+ */
 export function scenario (name, type, {css, screenshot} = {}) {
   const contextCopy = currentContext
   if (names.some(([existingName, existingContext]) =>
@@ -30,22 +46,26 @@ export function scenario (name, type, {css, screenshot} = {}) {
   })
 }
 
+/**
+ * Recieves the name of context and a function.
+ * Any scenarios created inside the function would have that context.
+ * @param {String} contextName
+ * @param {Function} func
+ */
 export function context (contextName, func) {
   currentContext = contextName
   func()
   currentContext = null
 }
 
-const Testshot = React.createClass({
-  propTypes: {
-    data: PropTypes.array,
-    server: PropTypes.shape({
-      host: PropTypes.string,
-      port: PropTypes.string
-    }),
-    routeData: PropTypes.object
-  },
-
+/**
+ * UI of Tessereact.
+ * @extends React.Component
+ * @property {String} props.host - host of the Tessereact server
+ * @property {String} props.port - port of the Tessereact server
+ * @property {RouteData} props.routeData
+ */
+class Testshot extends React.Component {
   render () {
     if (!data.length) {
       // TODO: Replace with nice and stylish welcome page :)
@@ -62,6 +82,16 @@ const Testshot = React.createClass({
 
     return <TestshotWindow host={this.props.server.host} port={this.props.server.port} data={data} routeData={this.props.routeData} />
   }
-})
+}
+
+if (PropTypes) {
+  Testshot.propTypes = {
+    server: PropTypes.shape({
+      host: PropTypes.string.isRequired,
+      port: PropTypes.string.isRequired
+    }),
+    routeData: PropTypes.object.isRequired
+  }
+}
 
 export default Testshot
