@@ -35,8 +35,8 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 /**
  * Convert scenario array to the tree consumed by Navigation or List element.
  *
- * @param {Array<Scenario>} scenarios
- * @returns {Array<Context|Scenario>} tree of context and scenarios
+ * @param {Array<ScenarioObject>} scenarios
+ * @returns {Array<ContextObject|ScenarioObject>} tree of contexts and scenarios
  */
 function generateTreeNodes(scenarios) {
   var groupedByContext = (0, _groupBy3.default)(scenarios, 'context');
