@@ -20,7 +20,7 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
  * In that case, redirect to the correct route.
  *
  * @param {RouteData} routeData
- * @param {Array<Scenario>} scenarios
+ * @param {Array<ScenarioObject>} scenarios
  */
 function checkIfRouteExists(routeData, scenarios) {
   var _routeData$params = routeData.params,
@@ -49,7 +49,7 @@ function checkIfRouteExists(routeData, scenarios) {
  * Check if the current route is home (/) and redirect to the first failing scenario.
  *
  * @param {RouteData} routeData
- * @param {Array<Scenario>} scenarios
+ * @param {Array<ScenarioObject>} scenarios
  */
 function checkForHomeRoute(routeData, scenarios) {
   var routeName = routeData.route.name;
@@ -72,7 +72,7 @@ function redirectToFirstScenario(scenarios) {
  * In the case when none of the scenarios is failing, redirect to the first scenario.
  *
  * @param {RouteData} routeData
- * @param {Array<Scenario>} scenarios
+ * @param {Array<ScenarioObject>} scenarios
  */
 function redirectToFirstFailingScenario(scenarios) {
   var scenario = scenarios.find(function (s) {
