@@ -15,10 +15,10 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
 /**
  * Find a scenario by the name of the scenario and the name of its context.
  *
- * @param {Array<Scenario>} scenarios
+ * @param {Array<ScenarioObject>} scenarios
  * @param {String} contextName
  * @param {String} scenarioName
- * @returns {Scenario} wanted scenario
+ * @returns {ScenarioObject} wanted scenario
  */
 function findScenario(scenarios, contextName, scenarioName) {
   return scenarios.find(function (s) {
@@ -45,8 +45,8 @@ function findScenarioIndex(scenarios, contextName, scenarioName) {
  * move selected scenario(s) to the start of the array if there are any.
  *
  * @param {RouteData} routeData
- * @param {Array<Scenario>} scenarios
- * @returns {Array<Scenario>} scenario array prepared to be sent to server
+ * @param {Array<ScenarioObject>} scenarios
+ * @returns {Array<ScenarioObject>} scenario array prepared to be sent to server
  */
 function getScenariosToLoad(routeData, scenarios) {
   var _routeData$params = routeData.params,
@@ -88,9 +88,9 @@ function shiftContext(scenarios, contextName) {
 /**
  * Accept the scenario in the given scenario array and return the array.
  *
- * @param {Array<Scenario>} scenarios
- * @param {Scenario} acceptedScenario
- * @returns {Array<Scenario>} new scenario array
+ * @param {Array<ScenarioObject>} scenarios
+ * @param {ScenarioObject} acceptedScenario
+ * @returns {Array<ScenarioObject>} new scenario array
  */
 function acceptScenario(scenarios, acceptedScenario) {
   var scenarioIndex = findScenarioIndex(scenarios, acceptedScenario.context, acceptedScenario.name);
@@ -108,9 +108,9 @@ function acceptScenario(scenarios, acceptedScenario) {
  * Replace scenario with the new version sent by the server
  * in the given scenario array and return the array.
  *
- * @param {Array<Scenario>} scenarios
- * @param {Scenario} scenario - scenario sent by the server
- * @returns {Array<Scenario>} new scenario array
+ * @param {Array<ScenarioObject>} scenarios
+ * @param {ScenarioObject} scenario - scenario sent by the server
+ * @returns {Array<ScenarioObject>} new scenario array
  */
 function resolveScenario(scenarios, scenario) {
   var storedScenarioIndex = findScenarioIndex(scenarios, scenario.context, scenario.name);
@@ -135,10 +135,10 @@ function resolveScenario(scenarios, scenario) {
  * merge the result of the function with screenshotData
  * and return the resulting scenario array.
  *
- * @param {Array<Scenario>} scenarios
- * @param {Scenario} scenario
+ * @param {Array<ScenarioObject>} scenarios
+ * @param {ScenarioObject} scenario
  * @param {Function} callback
- * @returns {Array<Scenario>} new scenario array
+ * @returns {Array<ScenarioObject>} new scenario array
  */
 function changeScenarioScreenshotData(scenarios, scenario, callback) {
   var storedScenarioIndex = findScenarioIndex(scenarios, scenario.context, scenario.name);
@@ -154,8 +154,8 @@ function changeScenarioScreenshotData(scenarios, scenario, callback) {
  * Prepare scenario to be sent to server for acceptance:
  * return only those fields of scenario object which are required by the server.
  *
- * @param {Scenario} scenario
- * @returns {Scenario} scenario prepared to be sent to server
+ * @param {ScenarioObject} scenario
+ * @returns {ScenarioObject} scenario prepared to be sent to server
  */
 function requestScenarioAcceptance(scenario) {
   var payload = {
