@@ -20,6 +20,22 @@ function _possibleConstructorReturn(self, call) { if (!self) { throw new Referen
 
 function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
+var PropTypes = void 0;
+try {
+  PropTypes = require('prop-types');
+} catch (e) {}
+// Ignore optional peer dependency
+
+
+/**
+ * Component which renders a single-page application style link to a route.
+ * @extends React.Component
+ * @property {String} props.name - name of the route
+ * @property {Object} props.params - search params of the route
+ * @property {Node} props.children - children nodes of the element
+ * @property {RouteData} [props.className] - CSS class name of the resulting element
+ */
+
 var Link = function (_Component) {
   _inherits(Link, _Component);
 
@@ -52,10 +68,13 @@ var Link = function (_Component) {
   return Link;
 }(_react.Component);
 
-Link.propTypes = {
-  className: _react.PropTypes.string,
-  name: _react.PropTypes.string,
-  params: _react.PropTypes.object,
-  children: _react.PropTypes.node
-};
+if (PropTypes) {
+  Link.propTypes = {
+    className: PropTypes.string,
+    name: PropTypes.string,
+    params: PropTypes.object,
+    children: PropTypes.node
+  };
+}
+
 exports.default = Link;
