@@ -1,32 +1,36 @@
-import React, {PropTypes} from 'react'
+import React from 'react'
 import Context from '../Context'
 import Scenario from '../Scenario'
 import StyledList from '../../styled/List'
 
-const List = React.createClass({
-  propTypes: {
-    nodes: PropTypes.array,
-    child: PropTypes.bool,
-    selectedNode: PropTypes.object,
-    selectNode: PropTypes.func,
-    searchQuery: PropTypes.string
-  },
+let PropTypes
+try {
+  PropTypes = require('prop-types')
+} catch (e) {
+  // Ignore optional peer dependency
+}
 
+/**
+ * Component which represents the nodes tree in sidebar.
+ * @extends React.Component
+ * @property {Array<ContextObject|ScenarioObject>} props.nodes - tree of contexts and scenarios
+ * @property {String} [props.searchQuery]
+ * @property {Boolean} [props.child] - is the tree a subtree
+ */
+class List extends React.Component {
   _renderItem (node) {
     return node.children ? this._renderContext(node) : this._renderScenario(node)
-  },
+  }
 
   _renderContext (node) {
-    const {selectedNode, searchQuery, selectNode} = this.props
+    const {searchQuery} = this.props
 
     return <Context
       key={node.name}
       node={node}
-      selectedNode={selectedNode}
-      selectNode={selectNode}
       searchQuery={searchQuery}
     />
-  },
+  }
 
   _renderScenario (node) {
     const {searchQuery, child} = this.props
@@ -37,7 +41,7 @@ const List = React.createClass({
       searchQuery={searchQuery}
       child={child}
     />
-  },
+  }
 
   render () {
     return <StyledList>
@@ -48,10 +52,18 @@ const List = React.createClass({
             (Boolean(a.children) === Boolean(b.children) ? 0 : a.children ? -1 : 1) ||
               a.name.localeCompare(b.name)
           )
-          .map(this._renderItem)
+          .map(this._renderItem, this)
       }
     </StyledList>
   }
-})
+}
+
+if (PropTypes) {
+  List.propTypes = {
+    nodes: PropTypes.array.isRequired,
+    child: PropTypes.bool,
+    searchQuery: PropTypes.string
+  }
+}
 
 export default List
