@@ -1,17 +1,25 @@
-import React, { PropTypes } from 'react'
+import React from 'react'
 import { find } from 'lodash'
 
-const ScenarioView = React.createClass({
-  propTypes: {
-    data: PropTypes.array.isRequired,
-    routeData: PropTypes.object
-  },
+let PropTypes
+try {
+  PropTypes = require('prop-types')
+} catch (e) {
+  // Ignore optional peer dependency
+}
 
+/**
+ * Component which renders a single scenario in isolation mode.
+ * @extends React.Component
+ * @property {Array<ScenarioObject>} props.data - list of scenarios created by user
+ * @property {RouteData} props.routeData
+ */
+class ScenarioView extends React.Component {
   getInitialState () {
     return {
       element: null
     }
-  },
+  }
 
   componentWillMount () {
     const {
@@ -25,12 +33,19 @@ const ScenarioView = React.createClass({
     this.setState({
       element: scenario.getElement()
     })
-  },
+  }
 
   render () {
     const {element} = this.state
     return element
   }
-})
+}
+
+if (PropTypes) {
+  ScenarioView.propTypes = {
+    data: PropTypes.array.isRequired,
+    routeData: PropTypes.object.isRequired
+  }
+}
 
 export default ScenarioView
